+++ CHANGES.rst
@@ -7,6 +7,7 @@ in progress
 ===========
 - Use ``grafana-client-2.1.0``, remove monkeypatch
 - Tests: Improve fixture ``create_datasource`` to clean up afterwards
+- Tests: Add fixture ``create_dashboard`` to create dashboards at runtime
 
 
 2022-02-03 0.13.1